pyproject.toml
[project]
name = "thn-gastos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.4",
    "fastapi>=0.115.0",
    "langchain-core>=0.3.79",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "openai>=2.3.0",
    "pydantic>=2.12.1",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "uvicorn>=0.32.0",
]
